+++ src/constants.js
@@ -151,13 +151,16 @@ export const itemSortMap = {
 export const itemStatuses = [
   { label: 'ui-inventory.item.status.available', value: 'Available' },
   { label: 'ui-inventory.item.status.awaitingPickup', value: 'Awaiting pickup' },
+  { label: 'ui-inventory.item.status.awaitingDelivery', value: 'Awaiting delivery' },
   { label: 'ui-inventory.item.status.checkedOut', value: 'Checked out' },
   { label: 'ui-inventory.item.status.inProcess', value: 'In process' },
   { label: 'ui-inventory.item.status.inTransit', value: 'In transit' },
   { label: 'ui-inventory.item.status.missing', value: 'Missing' },
   { label: 'ui-inventory.item.status.onOrder', value: 'On order' },
   { label: 'ui-inventory.item.status.paged', value: 'Paged' },
-  { label: 'ui-inventory.item.status.withdrawn', value: 'Withdrawn' },
+  { label: 'ui-inventory.item.status.declaredLost', value: 'Declared lost' },
+  { label: 'ui-inventory.item.status.received', value: 'Received' },
+  { label: 'ui-inventory.item.status.orderClosed', value: 'Order closed' }
 ];
 
 export const segments = {