+++ src/settings/HRIDHendling/HRIDHandlingSettings.js
@@ -40,7 +40,7 @@ const validateStartWithMaxLength = value => validateFieldLength(value, START_WIT
 const validateAssignPrefixMaxLength = value => validateFieldLength(value, ASSIGN_PREFIX_MAX_LENGTH);
 
 @stripesConnect
-export class HRIDHandlingSettings extends Component {
+class HRIDHandlingSettings extends Component {
   static manifest = Object.freeze({
     hridSettings: {
       type: 'okapi',