+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Always retrieve fresh statistical-code-types in settings. Refs UIIN-949.
 * Replace deprecated babel dependencies. Refs UIIN-956, STCOR-381.
 * Add filter for modes of issuance. Refs UIIN-823.
+* Add "Volume" and "Year, caption" fields to the Instance record in a detailed view. Refs UIIN-954.
 
 ## [1.13.1](https://github.com/folio-org/ui-inventory/tree/v1.13.1) (2019-12-11)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v1.13.0...v1.13.1)