+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change history for ui-inventory
 
+## 1.13.4 (IN PROGRESS)
+
+* Add "Volume" and "Year, caption" fields to the Instance record in a detailed view. Refs UIIN-954.
+
 ## [1.13.3](https://github.com/folio-org/ui-inventory/tree/v1.13.3) (2020-01-03)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v1.13.2...v1.13.3)